로그인/로그아웃 리뷰 반영
- session key 상수화
- 로그인 관련 세션 데이터 입출력 책임 분리
