+++ src/main/java/com/social/instagram/controller/LoginController.java
@@ -2,7 +2,8 @@ package com.social.instagram.controller;
 
 import com.social.instagram.dto.LoginDto;
 import com.social.instagram.factory.EncryptionFactory;
-import com.social.instagram.service.LoginService;
+import com.social.instagram.service.AccountService;
+import com.social.instagram.service.SessionService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -13,29 +14,31 @@ import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 
 import static com.social.instagram.util.httpstatus.ResponseSuccessContents.RESPONSE_ENTITY_OK;
+import static com.social.instagram.util.session.SessionNameConstants.ACCOUNT;
 
 @RestController
 @RequiredArgsConstructor
 public class LoginController {
 
-    private final LoginService loginService;
     private final EncryptionFactory encryptionFactory;
+    private final SessionService sessionService;
+    private final AccountService accountService;
 
     @PostMapping("/login")
     public ResponseEntity<Void> accountLogin(@Valid @RequestBody LoginDto loginDto, HttpSession session) {
         String userId = loginDto.getUserId();
         String password = encryptionFactory.sha256Util().changeEncoding(loginDto.getPassword());
 
-        loginService.hasAccount(userId, password);
+        accountService.hasAccount(userId, password);
 
-        session.setAttribute("account", loginService.getAccount(userId));
+        session.setAttribute(ACCOUNT, sessionService.getAccount(userId));
 
         return RESPONSE_ENTITY_OK;
     }
 
     @PostMapping("/logout")
     public ResponseEntity<Void> accountLogout(HttpSession session) {
-        session.removeAttribute("account");
+        session.removeAttribute(ACCOUNT);
 
         return RESPONSE_ENTITY_OK;
     }
