+++ src/main/java/com/social/instagram/exception/DuplicateUserIdException.java
@@ -1,5 +1,5 @@
 package com.social.instagram.exception;
 
-public class DuplicateUserIdException extends IllegalArgumentException{
+public class DuplicateUserIdException extends IllegalArgumentException {
 
 }
