+++ src/main/java/com/social/instagram/service/AccountService.java
@@ -2,6 +2,7 @@ package com.social.instagram.service;
 
 import com.social.instagram.domain.Account;
 import com.social.instagram.exception.DuplicateUserIdException;
+import com.social.instagram.exception.NotUserDataException;
 import com.social.instagram.repository.AccountRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -22,6 +23,16 @@ public class AccountService {
         if (accountRepository.existsByUserId(userId)) {
             throw new DuplicateUserIdException();
         }
+    }
+
+    public void hasAccount(String userId, String password) {
+        if (!accountRepository.existsFindByUserIdAndPassword(userId, password)) {
+            throw new NotUserDataException();
+        }
+    }
+
+    public Account getAccount(String userId) {
+        return accountRepository.findByUserId(userId);
     }
 
 }
