+++ src/main/java/com/social/instagram/service/SessionService.java
@@ -1,0 +1,17 @@
+package com.social.instagram.service;
+
+import com.social.instagram.domain.Account;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class SessionService {
+
+    private final AccountService accountService;
+
+    public Account getAccount(String userId) {
+        return accountService.getAccount(userId);
+    }
+
+}
