+++ src/main/java/com/social/instagram/util/session/SessionNameConstants.java
@@ -1,0 +1,7 @@
+package com.social.instagram.util.session;
+
+public class SessionNameConstants {
+
+    public static final String ACCOUNT = "account";
+
+}
